requirements.txt
flask==2.3.3
cohere==4.37
numpy==1.24.3
PyPDF2==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.3.7 